requirements.txt
twilio==9.3.1
flask
azure-communication-callautomation==1.2.0